+++ features/step_definitions/stack_steps.rb
@@ -65,6 +65,10 @@ Given(/^I stub CloudFormation validate calls to fail validation with message "([
   allow(StackMaster.cloud_formation_driver).to receive(:validate_template).and_raise(Aws::CloudFormation::Errors::ValidationError.new('', message))
 end
 
+Given(/^I stub the CloudFormation driver$/) do
+  allow(StackMaster.cloud_formation_driver.class).to receive(:new).and_return(StackMaster.cloud_formation_driver)
+end
+
 When(/^an S3 file in bucket "([^"]*)" with key "([^"]*)" exists with content:$/) do |bucket, key, body|
   file = StackMaster.s3_driver.find_file(bucket: bucket, object_key: key)
   expect(file).to eq body